Fixed a package/dependency removal bug that has bugged me for months.
If you have the following scenario (where -> means transitive closure):

have: a b c -> a b c d
want: a b   -> a b   d

The previous version would try to delete 'd' and fail because it was a
valid dependency of a or b.

This involved a number of refactorings including adding:

  + map-to-package-deps
  + flatten (maybe I should try to hide these in a letrec?)
  + package-installed-with-deps
  + package-manifest-with-deps
  + symbol< and symbol-list<
  + topological-sort

Wrote a fairly full test suite to prevent regressions in the future.